Require SUPPORTED_NETWORKS in LendProvider
